requirements.txt
networkx
matplotlib
neo4j
openai
pandas
noderag
langchain
langchain-openai
python-dotenv
jupyter
chromadb
faiss-cpu
pytest
